requirements.txt
django<3.1
psycopg2==2.7.4
psycopg2-binary
django-filter==2.2.0
djangorestframework==3.10.3
Pillow
beautifulsoup4
requests